requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
fastapi==0.115.2
h11==0.14.0
idna==3.7
plumbum==1.8.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyOpenSSL==24.2.1
python-decouple==3.7
python-dotenv==0.21.1
rpyc==6.0.0
sniffio==1.3.0
starlette==0.40.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1
websockets==12.0
requests==2.32.3
PyYAML==6.0.2